+++ tilings/strategies/pointing.py
@@ -3,17 +3,20 @@ The directionless point placement strategy that is counted
 by the 'pointing' constructor.
 """
 from collections import Counter
-from typing import Callable, Dict, FrozenSet, Iterator, Optional, Tuple
+from itertools import product
+from typing import Callable, Dict, FrozenSet, Iterator, List, Optional, Tuple
 
 from comb_spec_searcher import Strategy
 from comb_spec_searcher.exception import StrategyDoesNotApply
 from comb_spec_searcher.strategies.constructor.disjoint import DisjointUnion
+from comb_spec_searcher.strategies.rule import Rule
 from comb_spec_searcher.strategies.strategy import StrategyFactory
 from comb_spec_searcher.typing import CombinatorialClassType, SubTerms, Terms
 from permuta.misc import DIR_NONE
 from tilings import GriddedPerm, Tiling
 from tilings.algorithms import RequirementPlacement
 from tilings.assumptions import ComponentAssumption, TrackingAssumption
+from tilings.strategies.assumption_insertion import AddAssumptionsStrategy
 from tilings.strategies.obstruction_inferral import ObstructionInferralStrategy
 from tilings.tiling import Cell
 
@@ -260,7 +263,7 @@ class AssumptionPointingStrategy(PointingStrategy):
         raise StrategyDoesNotApply("The assumption is just point cells!")
 
     def formal_step(self) -> str:
-        return super().formal_step() + f" in cells {set(self.cells)}~~~"
+        return super().formal_step() + f" in cells {set(self.cells)}"
 
     def constructor(
         self,
@@ -336,4 +339,115 @@ class AssumptionPointingFactory(StrategyFactory[Tiling]):
 
     @classmethod
     def from_dict(cls, d: dict) -> "AssumptionPointingFactory":
+        return cls()
+
+
+class RequirementPointingStrategy(AssumptionPointingStrategy):
+    def __init__(
+        self,
+        gps: Tuple[GriddedPerm, ...],
+        indices: Tuple[int, ...],
+        ignore_parent: bool = False,
+        inferrable: bool = True,
+        possibly_empty: bool = True,
+        workable: bool = True,
+    ):
+        assert len(gps) == len(indices)
+        self.gps = gps
+        self.indices = indices
+        self.cells = frozenset(gp.pos[idx] for gp, idx in zip(gps, indices))
+        self.assumption = TrackingAssumption.from_cells(self.cells)
+
+        super().__init__(
+            self.assumption, ignore_parent, inferrable, possibly_empty, workable
+        )
+
+    def decomposition_function(self, comb_class: Tiling) -> Tuple[Tiling, ...]:
+        if self.assumption not in comb_class.assumptions:
+            raise StrategyDoesNotApply("The assumption is not on tiling")
+        cells = self.cells_to_place(comb_class)
+        algo = RequirementPlacement(comb_class)
+        if cells:
+            return (
+                comb_class.add_obstructions(
+                    [GriddedPerm.point_perm(cell) for cell in cells]
+                ),
+            ) + algo.place_point_of_req(
+                self.gps, self.indices, DIR_NONE, include_not=True
+            )
+        raise StrategyDoesNotApply("The assumption is just point cells!")
+
+    def formal_step(self) -> str:
+        return super().formal_step() + f" in {self.gps} at indices {self.indices}"
+
+    def extra_parameters(
+        self, comb_class: Tiling, children: Optional[Tuple[Tiling, ...]] = None
+    ) -> Tuple[Dict[str, str], ...]:
+        if children is None:
+            children = self.decomposition_function(comb_class)
+        empty_params = ObstructionInferralStrategy(
+            [GriddedPerm.point_perm(cell) for cell in self.cells_to_place(comb_class)]
+        ).extra_parameters(comb_class)
+        even_index = tuple(
+            child for idx, child in enumerate(children[1:]) if idx % 2 == 0
+        )
+        odd_index = tuple(
+            child for idx, child in enumerate(children[1:]) if idx % 2 == 1
+        )
+        res: List[Optional[Dict[str, str]]] = [None for _ in children[1:]]
+        for idx, param in enumerate(
+            PointingStrategy.extra_parameters(self, comb_class, even_index)
+        ):
+            res[2 * idx] = param
+        for idx, param in enumerate(
+            PointingStrategy.extra_parameters(self, comb_class, odd_index)
+        ):
+            res[2 * idx + 1] = param
+        return empty_params + tuple(res)
+
+    def to_jsonable(self) -> dict:
+        d = super().to_jsonable()
+        d["gps"] = [gp.to_jsonable() for gp in self.gps]
+        d["indices"] = self.indices
+        d.pop("assumption")
+        return d
+
+    @classmethod
+    def from_dict(cls, d: dict) -> "AssumptionPointingStrategy":
+        return cls(
+            tuple(GriddedPerm.from_dict(gp) for gp in d.pop("gps")),
+            tuple(d.pop("indices")),
+            **d,
+        )
+
+    def __repr__(self) -> str:
+        return (
+            self.__class__.__name__
+            + f"({self.gps}, {self.indices}, {self.ignore_parent}, "
+            f"{self.inferrable}, {self.possibly_empty}, {self.workable})"
+        )
+
+
+class RequirementPointingFactory(StrategyFactory[Tiling]):
+    def __call__(self, comb_class: Tiling) -> Iterator[Rule]:
+        for gp, indices in product(
+            Tiling(comb_class.obstructions).gridded_perms_of_length(2), [(0,), (1,)]
+        ):
+            gps = (gp,)
+            strategy = RequirementPointingStrategy(gps, indices)
+            parent = comb_class
+            if strategy.assumption not in comb_class.assumptions:
+                rule = AddAssumptionsStrategy([strategy.assumption])(comb_class)
+                yield rule
+                parent = rule.children[0]
+            yield strategy(parent)
+
+    def __str__(self) -> str:
+        return "requirement pointing strategy"
+
+    def __repr__(self) -> str:
+        return self.__class__.__name__ + "()"
+
+    @classmethod
+    def from_dict(cls, d: dict) -> "RequirementPointingFactory":
         return cls()