+++ CHANGELOG.md
@@ -43,9 +43,9 @@ swapped around a fusable row or column.
 - `FactorWithInterleavingStrategy.backward_map` so you can now generate permutation 
   from specifications using interleaving factors.
 - `DummyStrategy` that gives a quick template for making strategies. 
-- `PointingStrategy` and `AssumptionPointingFactory` that place points directionless 
-  in non-point cells. This are a non-productive strategy so should be used with 
-  `RuleDBForest`. 
+- `PointingStrategy`, `AssumptionPointingFactory` and `RequirementPointingFactory` 
+  that place points directionless in non-point cells. This are a non-productive 
+  strategy so should be used with `RuleDBForest`. 
 - `UnfusionFactory` that unfuses either all the rows or columns. Also non-productive.
 - `FusableRowAndColumnPlacementFactory` places fusable rows and columns.